+++ R/tbl_stack.R
@@ -3,7 +3,8 @@
 #' Assists in patching together more complex tables. `tbl_stack()` appends two
 #' or more `tbl_regression` or `tbl_merge` objects.
 #' {gt} attributes from the first regression object are utilized for output table.
-#' If combining `tbl_stack()` and `tbl_merge()`, merge first then stack.
+#' If combining `tbl_stack()` and `tbl_merge()`, merge first then stack, or stack
+#' first and then merge.
 #'
 #' @param tbls list of gtsummary regression objects
 #' @family tbl_regression tools
@@ -15,14 +16,18 @@
 #' # Example 1 - stacking two tbl_regression objects
 #' t1 <-
 #'   glm(response ~ trt, trial, family = binomial) %>%
-#'   tbl_regression(exponentiate = TRUE,
-#'                  label = list("trt" ~ "Treatment (unadjusted)"))
+#'   tbl_regression(
+#'     exponentiate = TRUE,
+#'     label = list("trt" ~ "Treatment (unadjusted)")
+#'   )
 #'
 #' t2 <-
 #'   glm(response ~ trt + grade + stage + marker, trial, family = binomial) %>%
-#'   tbl_regression(include = "trt",
-#'                  exponentiate = TRUE,
-#'                  label = list("trt" ~ "Treatment (adjusted)"))
+#'   tbl_regression(
+#'     include = "trt",
+#'     exponentiate = TRUE,
+#'     label = list("trt" ~ "Treatment (adjusted)")
+#'   )
 #'
 #' tbl_stack_ex1 <- tbl_stack(list(t1, t2))
 #'
@@ -30,14 +35,18 @@
 #' library(survival)
 #' t3 <-
 #'   coxph(Surv(ttdeath, death) ~ trt, trial) %>%
-#'   tbl_regression(exponentiate = TRUE,
-#'                  label = list("trt" ~ "Treatment (unadjusted)"))
+#'   tbl_regression(
+#'     exponentiate = TRUE,
+#'     label = list("trt" ~ "Treatment (unadjusted)")
+#'   )
 #'
 #' t4 <-
 #'   coxph(Surv(ttdeath, death) ~ trt + grade + stage + marker, trial) %>%
-#'   tbl_regression(include = "trt",
-#'                  exponentiate = TRUE,
-#'                  label = list("trt" ~ "Treatment (adjusted)"))
+#'   tbl_regression(
+#'     include = "trt",
+#'     exponentiate = TRUE,
+#'     label = list("trt" ~ "Treatment (adjusted)")
+#'   )
 #'
 #'
 #' # first merging, then stacking
@@ -46,8 +55,10 @@
 #' tbl_stack_ex2 <-
 #'   tbl_stack(list(row1, row2)) %>%
 #'   as_gt() %>%
-#'   tab_footnote(footnote = "Adjusted for cancer grade, state, and marker level.",
-#'                locations = cells_data(columns = "label", rows = 4))
+#'   tab_footnote(
+#'     footnote = "Adjusted for cancer grade, state, and marker level.",
+#'     locations = cells_data(columns = "label", rows = 4)
+#'   )
 #' @section Example Output:
 #' \if{html}{Example 1}
 #'
@@ -73,6 +84,11 @@ tbl_stack <- function(tbls) {
   tbls_length <- length(tbls)
   if (tbls_length < 2L) stop("Supply 2 or more gtsummary regression objects to 'tbls ='")
 
+  # checking if there are multiple input types
+  if (map_chr(tbls, class) %>% unique() %>% length() > 1) {
+    message("Multiple gtsummary object classes detected. Displayed results default to first input class type.")
+  }
+
   # stacking tables ------------------------------------------------------------
   results <- tbls[[1]][names(tbls[[1]]) %>% intersect(c("inputs", "gt_calls", "estimate_funs", "pvalue_funs", "qvalue_funs"))]
 
@@ -84,7 +100,7 @@ tbl_stack <- function(tbls) {
 
   # returning results ----------------------------------------------------------
   results$call_list <- list(tbl_stack = match.call())
-  results$tbl_regression_list = tbls
+  results$tbl_regression_list <- tbls
 
   class(results) <- "tbl_stack"
   return(results)