+++ R/tbl_uvregression.R
@@ -133,7 +133,9 @@ tbl_uvregression <- function(data, method, y, method.args = NULL,
   # get all x vars
   x_vars <- names(data) %>%
     setdiff( # removing outcome variable(s)
-      paste0(y, "~1") %>% stats::as.formula() %>% all.vars()
+      paste0(y, "~1") %>%
+        stats::as.formula() %>%
+        all.vars()
     ) %>%
     setdiff( # removing potential variables added to model formula (e.g. random intercepts)
       all.vars(stats::as.formula(formula)[[3]]) %>% remove_one_x() # the one x removed is the {x}
@@ -266,7 +268,7 @@ gt_tbl_uvregression <- quote(list(
   # indenting levels and missing rows
   tab_style_text_indent = glue(
     "tab_style(",
-    "style = cells_styles(text_indent = px(10), text_align = 'left'),",
+    "style = cell_text(indent = px(10), align = 'left'),",
     "locations = cells_data(",
     "columns = vars(label),",
     "rows = row_type != 'label'",